requirements.txt
requests==2.32.3
PyYAML==6.0.2
pytest==8.3.5
pytest-mock==3.14.0
